#12 프로필 사진 코드를 역할에 맞게 분리(UI, 데이터)
components/ProfilePhoto.js: UI에 관련된 코드
container/ProfilePhotoContainer.js: 데이터베이스 및 스토리지에 사진 정보 저장 및 요청하여 받아오는 코드 작성
components/ProfilePhoto.css: 기본 CSS작업
