+++ src/components/Profile/ProfilePhoto.css
@@ -1,13 +1,19 @@
 .profilePhoto {
-  position: relative;
   width: 200px;
   margin: auto;
 }
 .fileUploadBtn {
+  position: relative;
+}
+.fileUploadBtn > span {
+  cursor: pointer;
+}
+.fileUploadBtn > input {
   position: absolute;
   top: 0;
   left: 0;
   width: 100%;
   height: 100%;
   opacity: 0;
+  cursor: pointer;
 }
